requirements.txt
numpy>=1.26.4
scipy>=1.13.1
pandas>=2.2.3
matplotlib>=3.9.2
seaborn>=0.13.2
scikit-learn>=1.6.1
Pillow>=11.3.0
